reporting/requirements.txt
matplotlib>=3.8
numpy>=1.26
pandas>=2.1
reportlab>=4.0
python-pptx>=0.6.23
Pillow>=10.0
anthropic>=0.39.0
